+++ CHANGELOG.md
@@ -7,7 +7,8 @@
   exceptions directly rather than wrapping in a `WrapperRuntimeException` now that the underlying `DataFetcher`
   supplied by the `graphql-java` library throws an exception.
 * Remove `WrapperRuntimeException` now that it is unused.
-* Remove `AbstractDomgenGraphQLEndpoint.wrapDataFetcherError(...)` as it is unused.
+* Remove unused methods from `AbstractDomgenGraphQLEndpoint` which includes; `wrapDataFetcherError(...)`,
+  `getMessage(...)`, `throwableToError(...)`, `unwrap(...)` and `isWrapperException(...)`.
 
 ### [v1.3.0](https://github.com/realityforge/graphql-domgen-support/tree/v1.3.0) (2019-05-10)
 [Full Changelog](https://github.com/realityforge/graphql-domgen-support/compare/v1.2.0...v1.3.0)